+++ e2e/playground-vue.spec.ts
@@ -23,7 +23,7 @@ const SECTIONS = [
     testId: 'event-calendar-timeline-alternating',
     name: 'event-calendar-timeline-alternating.png',
   },
-  { testId: 'event-calendar-timeline-road', name: 'event-calendar-timeline-road.png' },
+  { testId: 'event-calendar-timeline-roadmap', name: 'event-calendar-timeline-roadmap.png' },
   { testId: 'selection-rules', name: 'selection-rules.png' },
   { testId: 'holidays', name: 'holidays.png' },
   { testId: 'holidays-and-rules', name: 'holidays-and-rules.png' },